Add `stream-name:*` syntax to DynamoJournal.find()
In most cases you only ever read or write a single Aggregate at a time,
for which you only need a single stream. In some cases, however, it's
useful to have access not only to a single stream, but to all the
streams of a related type.

This commit also fixes all our tests