Implement __UnauthorizedError event type and JWT authentication
If a websocket connection is denied with HTTP 401/403 during the
handshake, the client closes the connection but there's no way for
the client side JS to see the reason the connection closed, it
looks exactly like a network error.

This commit allows any connection, but checks the JWT in the
__Subscribe command. If authorization fails, it sends an
__UnauthorizedError event instead so that the client knows what
caused the error.